+++ packages/_private/e2e-demo-test/src/helpers.ts
@@ -1,4 +1,4 @@
-import { Locator } from '@playwright/test';
+import { Locator, Page } from '@playwright/test';
 import type {
   SynCheckbox,
   SynInput,
@@ -30,6 +30,9 @@ export const fillField = async (locator: Locator, value: string, selector = 'inp
   await locator.focus();
   await locator.locator(selector).fill(value);
   if (blur) {
+    // This is needed for firefox to recognize that the shadow root input has been left
+    // Otherwise, it will not trigger the syn-change event
+    await locator.locator(selector).blur();
     await locator.blur();
   }
 };
@@ -68,3 +71,68 @@ export const getFrameworkPort = (frameworkName: AvailableFrameworks) => {
   }
   return framework.port;
 };
+
+/**
+ * Wait for an event to be fired on the page
+ * @param page The page the event should be bound to
+ * @param event The event to use
+ * @param timeout The timeout to use
+ * @param eventOptions The event options to use
+ * @returns The page evaluation object for playwright
+ */
+export const waitForEvent = <T extends Event>(
+  page: Page,
+  event: keyof DocumentEventMap,
+  timeout = 500,
+  eventOptions: AddEventListenerOptions = { once: true },
+) => page.evaluate(({
+  event: playwrightEvent,
+  eventOptions: playwrightEventOptions,
+  timeout: playwrightTimeout,
+}) => new Promise<T>((resolve, reject) => {
+  const listener = ((e: T) => resolve(e)) as EventListener;
+  document.addEventListener(playwrightEvent, listener, playwrightEventOptions);
+
+  setTimeout(() => {
+    document.removeEventListener(playwrightEvent, listener, playwrightEventOptions);
+    reject(new Error(`Timeout waiting for event: ${playwrightEvent}`));
+  }, playwrightTimeout);
+}), {
+  event,
+  eventOptions,
+  timeout,
+});
+
+/**
+ * Do wait for an event to be fired on the page
+ * @param page The page the event should be bound to
+ * @param event The event to use
+ * @param timeout The timeout to use
+ * @param eventOptions The event options to use
+ * @returns The page evaluation object for playwright
+ */
+export const hasEvent = <T extends Event>(
+  page: Page,
+  event: keyof DocumentEventMap,
+  timeout = 500,
+  eventOptions: AddEventListenerOptions = { once: true },
+) => waitForEvent<T>(page, event, timeout, eventOptions)
+  .then(() => true)
+  .catch(() => false);
+
+/**
+ * Do NOT wait for an event to be fired on the page
+ * @param page The page the event should be bound to
+ * @param event The event to use
+ * @param timeout The timeout to use
+ * @param eventOptions The event options to use
+ * @returns The page evaluation object for playwright
+ */
+export const hasNoEvent = <T extends Event>(
+  page: Page,
+  event: keyof DocumentEventMap,
+  timeout = 500,
+  eventOptions: AddEventListenerOptions = { once: true },
+) => waitForEvent<T>(page, event, timeout, eventOptions)
+  .then(() => false)
+  .catch(() => true);