+++ packages/_private/e2e-demo-test/src/test.selector.ts
@@ -18,6 +18,15 @@ const AllComponentSelectors = {
   comboboxFormReset: '#tab-content-Combobox form syn-button',
   comboboxLink: '#tab-Combobox',
 
+  // Input
+  input417NoValue: '#tab-content-Input syn-input[data-testid="input-417-numeric-no-value"]',
+  input417NumericMinFractionDigitsModern: '#tab-content-Input syn-input[data-testid="input-838-numeric-modern-min-fraction-digits"]',
+  input417NumericMinFractionDigitsNative: '#tab-content-Input syn-input[data-testid="input-838-numeric-native-min-fraction-digits"]',
+  input417NumericModern: '#tab-content-Input syn-input[data-testid="input-417-numeric-modern"]',
+  input417NumericNative: '#tab-content-Input syn-input[data-testid="input-417-numeric-native"]',
+  inputContent: '#tab-content-Input',
+  inputLink: '#tab-Input',
+
   // NgModelUpdateOn
   ngModelUpdateOnCheckbox: '#tab-content-NgModelUpdateOn syn-checkbox[data-testid="checkbox-808"]',
   ngModelUpdateOnCheckboxChanged: '#tab-content-NgModelUpdateOn syn-checkbox[data-testid="checkbox-changed-808"]',