+++ packages/components/scripts/vendorism/custom/input.vendorism.js
@@ -1,4 +1,9 @@
-import { addSectionAfter, addSectionBefore, replaceSections } from '../replace-section.js';
+import {
+  addSectionAfter,
+  addSectionBefore,
+  addSectionsAfter,
+  replaceSections,
+} from '../replace-section.js';
 import { removeSections } from '../remove-section.js';
 
 const FILES_TO_TRANSFORM = [
@@ -103,11 +108,11 @@ export`,
       return true;
     }
 
-    if (this.min === undefined || this.min === null || this.disabled) {
+    if (this.min === undefined || this.min === null) {
       return false;
     }
 
-    const min = typeof this.min === 'string' ? parseFloat(this.min) : this.min
+    const min = typeof this.min === 'string' ? parseFloat(this.min) : this.min;
     return this.valueAsNumber <= min;
   }
 
@@ -120,7 +125,7 @@ export`,
       return false;
     }
 
-    const max = typeof this.max === 'string' ? parseFloat(this.max) : this.max
+    const max = typeof this.max === 'string' ? parseFloat(this.max) : this.max;
     return this.valueAsNumber >= max;
   }`,
   );
@@ -186,6 +191,363 @@ export`,
   content = content.replace(
     "@property() title = '';",
     "@property({ reflect: true }) title = '';",
+  );
+
+  // #417: Numeric Strategies
+  content = addSectionsAfter([
+    // Needed imports
+    [
+      "import type { SynergyFormControl } from '../../internal/synergy-element.js';",
+      `import {
+  type NumericStrategy,
+  createNumericStrategy,
+  nativeNumericStrategy,
+  modernNumericStrategy,
+} from './strategies.js';
+import { formatNumber } from './formatter.js';
+import type { SynClampDetails } from '../../events/syn-clamp.js';`,
+    ],
+    // Event documentation
+    [
+      "@event syn-invalid - Emitted when the form control has been checked for validity and its constraints aren't satisfied.",
+      ' * @event syn-clamp - Emitted if the numeric strategy allows autoClamp and the value is clamped to the min or max attribute.',
+    ],
+    // Add the numeric strategy property
+    [
+      "@property() inputmode: 'none' | 'text' | 'decimal' | 'numeric' | 'tel' | 'search' | 'email' | 'url';",
+      `
+  /**
+   * Optional options that should be passed to the \`NumberFormatter\` when formatting the value.
+   * This is used to format the number when the input type is \`number\`.
+   * Note this can only be set via \`property\`, not as an \`attribute\`!
+   */
+  @property({
+    attribute: false,
+    reflect: false,
+    type: Object,
+  }) numberFormatterOptions: Intl.NumberFormatOptions;
+
+  /**
+   * The minimal amount of fraction digits to use for numeric values.
+   * Used to format the number when the input type is \`number\`.
+   */
+  @property({
+    attribute: 'min-fraction-digits',
+    type: Number,
+  }) minFractionDigits: number;
+
+  /**
+   * The maximal amount of fraction digits to use for numeric values.
+   * Used to format the number when the input type is \`number\`.
+   */
+  @property({
+    attribute: 'max-fraction-digits',
+    type: Number,
+  }) maxFractionDigits: number;
+
+  #numericStrategy: NumericStrategy = nativeNumericStrategy;
+
+  /**
+   * Defines the strategy for handling numbers in the numeric input.
+   * This is used to determine how the input behaves when the user interacts with it.
+   *
+   * Includes the following configuration options:
+   *
+   * - **autoClamp**: If true, the input will clamp the value to the min and max attributes.
+   * - **noStepAlign**: If true, the input will not align the value to the step attribute.
+   * - **noStepValidation**: If true, the input will not validate the value against the step attribute.
+   * 
+   * You may provide this as one of the following values:
+   *
+   * - 'native': Uses the native browser implementation.
+   * - 'modern': Uses a more intuitive implementation:
+   *   - Values are clamped to the nearest min or max value.
+   *   - Stepping is inclusive to the provided min and max values.
+   *   - Provided stepping is no longer used in validation.
+   * - An object that matches the \`NumericStrategy\` type. Note this can only be set via \`property\`, not as an \`attribute\`!
+   */
+  @property({
+    attribute: 'numeric-strategy',
+    converter: {
+      fromAttribute: (value) => {
+        return value === 'modern'
+          ? modernNumericStrategy
+          : nativeNumericStrategy;
+      },
+    },
+    type: Object,
+  })
+  set numericStrategy(value: 'native' | 'modern' | Partial<NumericStrategy>) {
+    switch (typeof value) {
+      case 'string': this.#numericStrategy = value === 'modern' ? modernNumericStrategy : nativeNumericStrategy; break;
+      case 'object': this.#numericStrategy = createNumericStrategy(value); break;
+      default: this.#numericStrategy = nativeNumericStrategy;
+    }
+  }
+
+  /**
+   * @default nativeNumericStrategy
+   * @todo: This must be changed to "modern" in Synergy@3
+   */
+  get numericStrategy(): 'native' | 'modern' | Partial<NumericStrategy> {
+    return this.#numericStrategy;
+  }
+      `,
+    ],
+    // Add the numeric strategy switch to handleChange
+    [
+      'private handleChange() {',
+      `    if (this.type === 'number' && (this.#isNumberFormattingEnabled() || this.#numericStrategy.autoClamp)) {
+      const { eventObj, shouldClamp, nextValue } = this.handleNumericStrategyAutoClamp();
+      let initialNextValue = this.#numericStrategy.autoClamp ? nextValue : this.valueAsNumber;
+
+      // Make sure to flag non numeric values as invalid
+      if (isNaN(initialNextValue)) {
+        // Make sure to set the value to the min or max value if the input is empty
+        // If neither min nor max are set, we set the value to 0
+        const { max, min } = this;
+
+        if (max !== undefined && max !== null) {
+          initialNextValue = typeof max === 'string' ? parseFloat(max) : +max;
+        } else if (min !== undefined && min !== null) {
+          initialNextValue = typeof min === 'string' ? parseFloat(min) : +min;
+        } else {
+          initialNextValue = 0;
+        }
+      }
+
+      this.value = this.#isNumberFormattingEnabled()
+        ? this.#formatNumber(initialNextValue)
+        : initialNextValue.toString();
+
+      // Make sure to wait for the updateComplete to be done before updating the validity
+      // and firing events. This is needed because valueAsNumber is not updated yet when the event is fired
+      this.updateComplete.then(() => {
+
+        if (shouldClamp && eventObj) {
+          this.emit('syn-clamp', eventObj);
+        }
+
+        this.formControlController.updateValidity();
+        this.emit('syn-change');
+      });
+      return;
+    }`,
+    ],
+    // Make sure to disable the sync of the step property if noStepValidation is used
+    [
+      'handleStepChange() {',
+      `    // If the numericStrategy has noStepValidation set, skip this as the inputs step will always set to "any".
+    if (this.#numericStrategy.noStepValidation) {
+      return;
+    }
+      `,
+    ],
+  ], content);
+  // /#417
+
+  // #818
+  content = addSectionsAfter([
+    // Handle keydown
+    [
+      'private handleKeyDown(event: KeyboardEvent) {',
+      `    if (this.#numericStrategy.noStepAlign && this.type === 'number') {
+      const { key } = event;
+      if (key === 'ArrowUp' || key === 'ArrowDown') {
+        event.preventDefault();
+        event.stopPropagation();
+
+        if (key === 'ArrowUp') {
+          this.handleStepUp();
+        } else if (key === 'ArrowDown') {
+          this.handleStepDown();
+        }
+        this.handleChange();
+        return;
+      }
+    }`,
+    ],
+    // Handle step up
+    [
+      'stepUp() {',
+      `    if (this.#numericStrategy.noStepAlign) {
+      const { max, step, valueAsNumber } = this;
+
+      // Needed because the input could be empty. In this case, valueAsNumber is NaN
+      const usedInitialValue = Number.isNaN(valueAsNumber) ? 0 : valueAsNumber;
+      const usedMin = typeof this.min === 'string' ? parseFloat(this.min) : this.min;
+      const usedMax = typeof max === 'string' ? parseFloat(max) : max;
+      const usedStep = (typeof step === 'undefined' || step === null || step === 'any') ? 1 : typeof step === 'number' ? step : parseFloat(step);
+
+      let wantedNextValue = usedInitialValue + usedStep;
+
+      if (typeof usedMax === 'number' && usedMax < wantedNextValue) {
+        wantedNextValue = usedMax;
+      } else if (typeof usedMin === 'number' && usedMin > wantedNextValue) {
+        wantedNextValue = usedMin;
+      }
+
+      const finalStringValue = this.#isNumberFormattingEnabled()
+        ? this.#formatNumber(wantedNextValue)
+        : wantedNextValue.toString();
+
+      this.input.value = finalStringValue;
+
+      if (this.value !== this.input.value) {
+        this.value = this.input.value;
+      }
+      return;
+    }
+`,
+    ],
+    // Handle step down
+    [
+      'stepDown() {',
+      `    if (this.#numericStrategy.noStepAlign) {
+      const { min, max, step, valueAsNumber } = this;
+
+      // Needed because the input could be empty. In this case, valueAsNumber is NaN
+      const usedInitialValue = Number.isNaN(valueAsNumber) ? 0 : valueAsNumber;
+      const usedMin = typeof min === 'string' ? parseFloat(min) : min;
+      const usedMax = typeof max === 'string' ? parseFloat(max) : max;
+      const usedStep = (typeof step === 'undefined' || step === null || step === 'any') ? 1 : typeof step === 'number' ? step : parseFloat(step);
+
+      let wantedNextValue = usedInitialValue - usedStep;
+      
+      if (typeof usedMin === 'number' && usedMin > wantedNextValue) {
+        wantedNextValue = usedMin;
+      } else if (typeof usedMax === 'number' && usedMax < wantedNextValue) {
+        wantedNextValue = usedMax;
+      }
+
+      const finalStringValue = this.#isNumberFormattingEnabled()
+        ? this.#formatNumber(wantedNextValue)
+        : wantedNextValue.toString();
+
+      this.input.value = finalStringValue;
+
+      if (this.value !== this.input.value) {
+        this.value = this.input.value;
+      }
+      return;
+    }
+`,
+    ],
+  ], content);
+
+  // Disable the stepper if the input uses noStepValidation
+  content = content.replace(
+    'ifDefined(this.step as number)',
+    "ifDefined(!this.#numericStrategy.noStepValidation ? this.step as number : 'any')",
+  );
+  // /#818
+
+  // Add the modern numeric strategy for autoclamp
+  content = addSectionBefore(
+    content,
+    'private handleChange() {',
+    `
+  private handleNumericStrategyAutoClamp() {
+    const {
+      valueAsNumber,
+      max: initialMax,
+      min: initialMin,
+    } = this;
+
+    // Exit early if there should be no clamping
+    if (!this.#numericStrategy.autoClamp) {
+      return {
+        eventObj: null,
+        shouldClamp: false,
+        nextValue: valueAsNumber,
+      }
+    }
+
+    const min = typeof initialMin === 'string' ? parseFloat(initialMin) : initialMin;
+    const max = typeof initialMax === 'string' ? parseFloat(initialMax) : initialMax;
+
+    let nextValue = valueAsNumber;
+    let clampEvent = '';
+    if (nextValue < min) {
+      nextValue = min;
+      clampEvent = 'min';
+    } else if (nextValue > max) {
+      nextValue = max;
+      clampEvent = 'max';
+    }
+
+    const eventObj = clampEvent ? {
+      detail: {
+        clampedTo: clampEvent as SynClampDetails['clampedTo'],
+        lastUserValue: valueAsNumber,
+      },
+    } : null;
+
+    return {
+      eventObj,
+      shouldClamp: !!eventObj,
+      nextValue,
+    };
+  }`,
+    {
+      newlinesAfterInsertion: 2,
+      tabsAfterInsertion: 1,
+    },
+  );
+  // /#417
+
+  // #838: Add formatter
+  content = addSectionBefore(
+    content,
+    'render() {',
+    `#formatNumber(value: number) {
+    return formatNumber(value, this.step, {
+      maximumFractionDigits: this.maxFractionDigits,
+      minimumFractionDigits: this.minFractionDigits,
+      ...this.numberFormatterOptions,
+    });
+  }`,
+    {
+      newlinesAfterInsertion: 2,
+      tabsAfterInsertion: 1,
+    },
+  );
+
+  content = addSectionBefore(
+    content,
+    'render() {',
+    `#isNumberFormattingEnabled() {
+    const {
+      numberFormatterOptions,
+      maxFractionDigits,
+      minFractionDigits,
+      step,
+    } = this;
+
+    const hasMaxFractionDigits = typeof maxFractionDigits !== 'undefined' && !Number.isNaN(maxFractionDigits);
+    const hasMinFractionDigits = typeof minFractionDigits !== 'undefined' && !Number.isNaN(minFractionDigits);
+
+    // Easy checks first: If we have a min or max fraction digit, proceed
+    if (hasMaxFractionDigits || hasMinFractionDigits) {
+      return true;
+    }
+    
+    // Check if there are any options that where provided via formatter options
+    if (typeof numberFormatterOptions === 'object') {
+      return true;
+    }
+
+    // As a last fallback, see if the step has a decimal value
+    // If it has, we should format according to the steps amount of fraction digits
+    const stepToUse = step === 'any' || !step ? 1 : +step;
+    const stepFractionDigits = stepToUse.toString().split('.')[1]?.length || 0;
+
+    return stepFractionDigits > 0;
+  }`,
+    {
+      newlinesAfterInsertion: 2,
+      tabsAfterInsertion: 1,
+    },
   );
 
   return {